% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDB.R
\name{createConnection}
\alias{createConnection}
\title{Create a DB connection for a remote or local data processing}
\usage{
createConnection(
  server = "localhost",
  database,
  password,
  user,
  dbms,
  port,
  pathToDriver
)
}
\arguments{
\item{server}{IP of the database server}

\item{database}{Name of the database}

\item{password}{Password of the database user}

\item{user}{Username of the database user}

\item{dbms}{The database management system}

\item{port}{Port of the database}

\item{pathToDriver}{Path to drivers' repository}
}
\description{
Create a DB connection for a remote or local data processing
}
