% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDB.R
\name{createConnectionSQLite}
\alias{createConnectionSQLite}
\title{Create a SQLite server for using when processing the data}
\usage{
createConnectionSQLite()
}
\description{
Create a SQLite server for using when processing the data
}
