% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectoryFunctions.R
\name{filterBySettings}
\alias{filterBySettings}
\title{Remove all instances for patient which occur after selected state occurrence}
\usage{
filterBySettings(connection, dbms, schema, settings)
}
\arguments{
\item{connection}{Connection to the database (DatabaseConnector)}

\item{dbms}{The database management system}

\item{schema}{Schema in the database where the tables are located}

\item{settings}{The settings of generating trajectories}
}
\description{
Remove all instances for patient which occur after selected state occurrence
}
\keyword{internal}
