% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preliminaryAnalysis.R
\name{getDistinctTrajectoriesTable}
\alias{getDistinctTrajectoriesTable}
\title{Create a table summarizing the total count of all distinct trajectories}
\usage{
getDistinctTrajectoriesTable(
  connection,
  dbms,
  schema,
  table = "patient_trajectories_combined"
)
}
\arguments{
\item{connection}{Connection to the database (DatabaseConnector)}

\item{dbms}{The database management system}

\item{schema}{Name of the used schema}

\item{table}{Name of the used table}
}
\description{
Create a table summarizing the total count of all distinct trajectories
}
