% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectoryFunctions.R
\name{getEdgesDataset}
\alias{getEdgesDataset}
\title{Query all data about edges and mean elapsed time}
\usage{
getEdgesDataset(connection, dbms, schema)
}
\arguments{
\item{connection}{Connection to the database (DatabaseConnector)}

\item{dbms}{The database management system}

\item{schema}{Schema in the database where the tables are located}
}
\description{
Query all data about edges and mean elapsed time
}
