% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectoryFunctions.R
\name{importTempData}
\alias{importTempData}
\title{Query all patients' data satifying the criteria}
\usage{
importTempData(connection, dbms, schema)
}
\arguments{
\item{connection}{Connection to the database (DatabaseConnector)}

\item{dbms}{The database management system}

\item{schema}{Schema in the database where the tables are located}

\item{trajectories}{The matches in the matching table}
}
\description{
Query all patients' data satifying the criteria
}
