% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectoryFunctions.R
\name{importTrajectoryData}
\alias{importTrajectoryData}
\title{Query all trajectories defined in the matching table as matches}
\usage{
importTrajectoryData(connection, dbms, schema, trajectories, settings)
}
\arguments{
\item{connection}{Connection to the database (DatabaseConnector)}

\item{dbms}{The database management system}

\item{schema}{Schema in the database where the tables are located}

\item{trajectories}{The matches in the matching table}

\item{settings}{The settings of generating trajectories}
}
\description{
Query all trajectories defined in the matching table as matches
}
