% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDB.R
\name{loadUITrajectories}
\alias{loadUITrajectories}
\title{Function for loading configuration of selected trajectories}
\usage{
loadUITrajectories(pathToFile = NULL, settings = NA)
}
\arguments{
\item{pathToFile}{The path to the settings file}
}
\description{
Function for loading configuration of selected trajectories
}
