% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectoryFunctions.R
\name{mergeNewSubjects}
\alias{mergeNewSubjects}
\title{Merge two tables, only updating subject_ids not present in tableA}
\usage{
mergeNewSubjects(connection, dbms, schema, tableA, tableB)
}
\arguments{
\item{connection}{Connection to the database (DatabaseConnector)}

\item{dbms}{The database management system}

\item{schema}{Schema in the database where the tables are located}

\item{tableA}{Source table (Source of new subject_ids)}
}
\description{
Merge two tables, only updating subject_ids not present in tableA
}
\keyword{internal}
