% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preliminaryAnalysis.R
\name{outputTrajectoryStatisticsTables}
\alias{outputTrajectoryStatisticsTables}
\title{Query data tables (matching, partially matching, not matching) defined in the settings for providing inclusion statistics}
\usage{
outputTrajectoryStatisticsTables(dataTable, settings = NULL)
}
\arguments{
\item{dataTable}{Result of getDistinctTrajectoriesTable function}

\item{settings}{The settings of generating trajectories}
}
\description{
Query data tables (matching, partially matching, not matching) defined in the settings for providing inclusion statistics
}
