% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectoryFunctions.R
\name{queryNodesDatasetGroup}
\alias{queryNodesDatasetGroup}
\title{Query data about nodes from specified group}
\usage{
queryNodesDatasetGroup(connection, dbms, schema, groupId = NULL)
}
\arguments{
\item{connection}{Connection to the database (DatabaseConnector)}

\item{dbms}{The database management system}

\item{schema}{Schema in the database where the tables are located}

\item{groupId}{Group id for patients}
}
\description{
Query data about nodes from specified group
}
