% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectoryFunctions.R
\name{removeAfterDatasetDB}
\alias{removeAfterDatasetDB}
\title{Remove all instances for patient which occur after selected state occurrence}
\usage{
removeAfterDatasetDB(connection, dbms, schema, selectedState, active = TRUE)
}
\arguments{
\item{connection}{Connection to the database (DatabaseConnector)}

\item{dbms}{The database management system}

\item{schema}{Schema in the database where the tables are located}

\item{selectedState}{Selected state}

\item{active}{Boolean indicating whether or not the function is used in TrajectoryMappings package}
}
\description{
Remove all instances for patient which occur after selected state occurrence
}
