% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preliminaryAnalysis.R
\name{transformData}
\alias{transformData}
\title{Prepare Cohort2Trajectory output to be used in the analyser}
\usage{
transformData(
  data,
  removeID = TRUE,
  removeSTART = TRUE,
  removeEXIT = TRUE,
  fixGender = TRUE,
  showOccurrance = TRUE
)
}
\arguments{
\item{data}{Trajectories dataset from Cohort2Trajectory}

\item{removeID}{If true the original id of subject will be lost}

\item{removeSTART}{If true all the "START" states will be removed}

\item{removeEXIT}{If true all the "EXIT" states will be removed}

\item{fixGender}{If true all the concept codes will be converted to "MALE", "FEMALE" or "OTHER"}

\item{showOccurrance}{If true all the states will have an ordinal indicating how many passings for a patient have occurred}
}
\description{
Prepare Cohort2Trajectory output to be used in the analyser
}
