% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDB.R
\name{createTrajectoriesTable}
\alias{createTrajectoriesTable}
\title{Create support tables and dump data into the database if it is not present}
\usage{
createTrajectoriesTable(
  connection,
  dbms,
  data = NULL,
  schema,
  tableName = NULL
)
}
\arguments{
\item{connection}{Connection to the database (DatabaseConnector)}

\item{dbms}{The database management system}

\item{data}{Imported data}

\item{schema}{Name of the used schema}

\item{tableName}{Name of the datatable if it exists in the database}
}
\description{
Create support tables and dump data into the database if it is not present
}
